// src/api/models/dto/UserInputDTO.ts
export interface UserInput {
     firstName: string;
     lastName: string;
     email: string;
     password: string;
     age: number;
     gender: string;
     profilePictureUrl: string;
     useOfCar: string;
     numberOfCars: number;
    //  teslUsername: string;
    //  teslaPassword: string;
}