// src/api/models/dto/UserInputForUserProfile.ts
export interface UserInputForUserProfile {
    email: string;
    userId: string;
    firstName: string;
    age: number;
    sex: string;
    phoneNumber: string;
    numberOfCars: 0;
    useOfCar: string;
    companyName: string;
    avatar: string;
    employeeId: string;
    distanceUnit: string;
    zoneId: string;
    currency: string;
  }