// src/api/models/dto/UserInputWithoutEmailPassDTO.ts
export interface UserInputWithoutEmailPassDTO {
     firstName: string;
     lastName: string;
     idToken: string;
     age: number;
     gender: string;
     profilePictureUrl: string;
     useOfCar: string;
     numberOfCars: number;
    //  teslUsername: string;
    //  teslaPassword: string;
}